change permissions management

+++ resources/views/dashboard/permissions/index.blade.php
@@ -1,6 +1,6 @@
 @extends('layouts.dashboard')
 
-@section('title', '| Users')
+@section('title', '| Permissions')
 
 @section('content')
     <div class="container-fluid">
@@ -8,34 +8,42 @@
             <div class="col-md-8 col-sm-12">
                 <h4 class="c-grey-900 mT-10 mB-30">
                     <i class="fa fa-key icon-left"></i>Available Permissions
-                    <a href="{{ route('permissions.create') }}" class="btn btn-outline-primary pull-right">
-                        <i class="fa fa-plus icon-left"></i>New Permission
-                    </a>
                 </h4>
                 <div class="bgc-white bd bdrs-3 p-20 mB-20">
                     <table id="dataTable" class="table table-bordered table-hover" cellspacing="0" width="100%">
                         <thead>
                             <tr>
-                                <th>Permission</th>
+                                <th>Name</th>
+                                <th>Display Name</th>
                                 <th>Operations</th>
                             </tr>
                         </thead>
+                        <tfoot>
+                            <tr>
+                                <th>Name</th>
+                                <th>Display Name</th>
+                                <th>Operations</th>
+                            </tr>
+                        </tfoot>
                         <tbody>
-                            @foreach  ($permissions as $permission)
+                            @foreach  ($permissions as $key => $permission)
                                 <tr>
-                                    <td>{{ $permission->name }}</td>           
+                                    <td>{{ $permission->name }}</td>
+                                    <td>{{ $permission->display_name }}</td>  
                                     <td>
                                         <div class="gap-5 peers">
                                             <div class="peer">
                                                 <a href="{{ route('permissions.edit', $permission->id) }}" class="btn btn-sm btn-outline-info">Edit</a>
                                             </div>
-                                            <div class="peer">
-                                                {!! Form::open(['method' => 'DELETE', 'route' => ['permissions.destroy', $permission->id] ]) !!}
-                                                    {!! Form::submit('Delete', ['class' => 'btn btn-sm btn-outline-danger']) !!}
-                                                 {!! Form::close() !!}
-                                            </div>
+                                            @if (!$permission->built_in)
+                                                <div class="peer">
+                                                    {!! Form::open(['method' => 'DELETE', 'route' => ['permissions.destroy', $permission->id] ]) !!}
+                                                        <button class="btn btn-sm btn-outline-danger" onclick="confirmDeletion(event)">Delete</button>
+                                                    {!! Form::close() !!}
+                                                </div>
+                                            @endif
                                         </div>
-                                    </td>
+                                    </td>         
                                 </tr>
                             @endforeach
                         </tbody>
@@ -45,3 +53,5 @@
         </div>
     </div>
 @endsection
+
+@include('layouts.utils.confirmDeletionScript')
